Use u32 throughout

+++ sparse_strips/vello_sparse_tests/tests/util.rs
@@ -335,7 +335,7 @@ pub(crate) fn check_ref(
     specific_name: &str,
     // Tolerance for pixel differences.
     threshold: u8,
-    diff_pixels: u16,
+    diff_pixels: u32,
     // Whether the test instance is the "gold standard" and should be used
     // for creating reference images.
     is_reference: bool,
@@ -367,7 +367,7 @@ pub(crate) fn check_ref(
         .into_rgba8();
     let actual = load_from_memory(&encoded_image).unwrap().into_rgba8();
 
-    let diff_result = get_diff(&ref_image, &actual, threshold, diff_pixels as u32);
+    let diff_result = get_diff(&ref_image, &actual, threshold, diff_pixels);
 
     if let Some((diff_image, diff_data)) = diff_result {
         if should_replace() && is_reference {
@@ -417,7 +417,7 @@ pub(crate) fn check_ref(
     specific_name: &str,
     // Tolerance for pixel differences.
     threshold: u8,
-    diff_pixels: u16,
+    diff_pixels: u32,
     // Must be `false` on `wasm32` as reference image cannot be written to filesystem.
     is_reference: bool,
     ref_data: &[u8],
@@ -430,7 +430,7 @@ pub(crate) fn check_ref(
 
     let ref_image = load_from_memory(ref_data).unwrap().into_rgba8();
 
-    let diff_image = get_diff(&ref_image, &actual, threshold, diff_pixels as u32);
+    let diff_image = get_diff(&ref_image, &actual, threshold, diff_pixels);
     if let Some((ref img, _)) = diff_image {
         append_diff_image_to_browser_document(specific_name, img);
         panic!("test didn't match reference image. Scroll to bottom of browser to view diff.");
